Begun work on the support for the stm l6470 driver.
What I am doing is adapting this: https://github.com/vanvught/rpidmx512/tree/master/lib-l6470.
It is OK to just copy it since the license is MIT. What is happening, I am switching the spi
interface to use the SPIDEV that is available on the RPI5. The library that was originaly used
is not available for the RPI5.